[70197] Refactor import sample data model
